requirements.txt
anyio==3.3.2
apturl==0.5.2
argon2-cffi==21.1.0
asn1crypto==0.24.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
Brlapi==0.6.6
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==6.7
colorama==0.4.4
command-not-found==0.3
contextvars==2.4
cryptography==2.1.4
cupshelpers==1.0
dataclasses==0.8
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
future==0.18.2
httplib2==0.9.2
idna==3.2
immutables==0.16
importlib-metadata==4.8.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.4
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
mistune==0.8.4
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.4.4
oauth==1.0.1
oauthlib==3.1.1
olefile==0.45.1
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
powerline-status==2.6
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.0.0
psutil==5.8.0
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.10.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.18.0
pyte==0.8.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.19.0
python-twitter==3.5
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
pyzmq==22.3.0
reportlab==3.4.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.13.2
six==1.16.0
sniffio==1.2.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
terminado==0.12.1
testpath==0.5.0
thefuck==3.31
tornado==6.1
traitlets==4.3.3
tweepy==4.0.1
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
xkit==0.0.0
zipp==3.5.1
zope.interface==4.3.2
